backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.3.2
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
google-generativeai==0.8.0
